requirements.txt
Django>=5.0.0
Pillow>=10.0.0
requests>=2.31.0
python-dateutil>=2.8.0
django-extensions>=3.2.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-filter>=23.0
django-tables2>=2.6.0
django-import-export>=3.3.0
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
